Execute machine specific Bash setup
This change brings in the ability to execute a custom Bash script
from the `.bashrc` file. The script that is executed is derived from
the machine name and found under `~/.config/bash`.

With this change, environment specific stuff can be achieved like
what I am doing on WSL, where I need access to the 1Password `op`
executable.

Signed-off-by: Wynand Pieterse <wpieterse2825@gmail.com>